requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
httpx==0.27.0
apscheduler==3.10.4
line-bot-sdk==3.11.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
pytest==8.3.0
pytest-asyncio==0.24.0
